Extract LoginWithQR types
