+++ test/components/views/settings/devices/LoginWithQR-test.tsx
@@ -20,7 +20,8 @@ import React from "react";
 import { MSC3906Rendezvous, RendezvousFailureReason } from "matrix-js-sdk/src/rendezvous";
 import { HTTPError, LoginTokenPostResponse } from "matrix-js-sdk/src/matrix";
 
-import LoginWithQR, { Click, Mode, Phase } from "../../../../../src/components/views/auth/LoginWithQR";
+import LoginWithQR from "../../../../../src/components/views/auth/LoginWithQR";
+import { Click, Mode, Phase } from "../../../../../src/components/views/auth/LoginWithQR-types";
 import type { MatrixClient } from "matrix-js-sdk/src/matrix";
 
 jest.mock("matrix-js-sdk/src/rendezvous");
