+++ test/components/views/settings/devices/LoginWithQRFlow-test.tsx
@@ -19,12 +19,8 @@ import React from "react";
 import { RendezvousFailureReason } from "matrix-js-sdk/src/rendezvous";
 
 import LoginWithQRFlow from "../../../../../src/components/views/auth/LoginWithQRFlow";
-import {
-    Click,
-    Phase,
-    LoginWithQRFailureReason,
-    FailureReason,
-} from "../../../../../src/components/views/auth/LoginWithQR";
+import { LoginWithQRFailureReason, FailureReason } from "../../../../../src/components/views/auth/LoginWithQR";
+import { Click, Phase } from "../../../../../src/components/views/auth/LoginWithQR-types";
 
 describe("<LoginWithQRFlow />", () => {
     const onClick = jest.fn();
